Refactoring of ConverterController
* removed AbstractConverterController due to uselessness  
* made `convertRate` not nullable  
* provided default values for `baseVerifier` and `targetVerifier`  
* a lot of cosmetic changes
