+++ converter/src/main/java/ru/touchin/converter/verifiers/IntegerVerifier.kt
@@ -3,11 +3,11 @@ package ru.touchin.converter.verifiers
 import ru.touchin.converter.commands.Command
 import java.math.BigDecimal
 
-class IntegerVerifier(maxIntegerNumbers: Int) : Verifier<BigDecimal> {
+class IntegerVerifier(maxIntegerNumbers: Int = 5) : Verifier<BigDecimal> {
 
-    private val pattern = "(\\d{1,$maxIntegerNumbers})((\\.)\\d*)?".toRegex()
+    private val pattern = "^\\d{1,$maxIntegerNumbers}$".toRegex()
 
-    override fun verify(text: String): Command<BigDecimal> = if (pattern.matches(text) == true) {
+    override fun verify(text: String): Command<BigDecimal> = if (pattern.matches(text)) {
         Command.Success()
     } else {
         Command.Fallback()
